PythonMicroserviceProject/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
aiosqlite
python-dotenv
loguru
python-dotenv
